new 1-10 sliders for FaceScanWidget
New sliders for FaceScanWidget that use 1 to 10 scale
Many bug fixes

+++ core/libs/facesengine/detection/facedetector.cpp
@@ -95,7 +95,7 @@ public:
 
         if (m_parameters.contains(QLatin1String("detectAccuracy")))
         {
-            backend()->setAccuracy(static_cast<float>(m_parameters.value(QLatin1String("detectAccuracy")).toDouble()));
+            backend()->setAccuracy(m_parameters.value(QLatin1String("detectAccuracy")).toInt());
         }
 
         if (m_parameters.contains(QLatin1String("detectSize")))
