+++ core/utilities/facemanagement/threads/facepipeline_p.h
@@ -102,7 +102,7 @@ Q_SIGNALS:
     friend class FacePipeline;
     void startProcess(const FacePipelineExtendedPackage::Ptr& package);
 
-    void accuracyAndModel(double detectAccuracy, FaceScanSettings::FaceDetectionModel detectModel, FaceScanSettings::FaceDetectionSize detectSize, double recognizeAccuracy, FaceScanSettings::FaceRecognitionModel recognizeModel);
+    void accuracyAndModel(int detectAccuracy, FaceScanSettings::FaceDetectionModel detectModel, FaceScanSettings::FaceDetectionSize detectSize, int recognizeAccuracy, FaceScanSettings::FaceRecognitionModel recognizeModel);
 
 private:
